🔖 Alpha version
- 🔨 Refactor
- ✨ Make url parameter optional
