+++ tests/test_detect.py
@@ -1,15 +1,13 @@
 import sys
 from typing import Dict
-from unittest import mock
 from unittest.mock import ANY
 
 import pytest
 
 import browsers
-from browsers import get_available_browsers
 
 """
-These tests assume that Chrome, Firefox, Safari and Microsoft Edge are installed. These are available in Github Actions.
+These tests are based on what browsers exists in Github Actions virtual environments.
 """
 
 
@@ -26,7 +24,7 @@ These tests assume that Chrome, Firefox, Safari and Microsoft Edge are installed
     ),
 )
 def test_get_available_browsers(browser: str) -> None:
-    available_browsers = dict(get_available_browsers())
+    available_browsers = dict(browsers.browsers())
     assert browser in available_browsers
 
 
@@ -137,35 +135,3 @@ def test_get_available_browsers(browser: str) -> None:
 )
 def test_get(browser: str, details: Dict) -> None:
     assert browsers.get(browser) == details
-
-
-@pytest.mark.parametrize(
-    "chrome_path",
-    (
-        pytest.param(
-            "/Applications/Google Chrome.app",
-            id="osx",
-            marks=pytest.mark.skipif(sys.platform != "darwin", reason="osx-only"),
-        ),
-        pytest.param(
-            "/usr/bin/google-chrome-stable",
-            id="linux",
-            marks=pytest.mark.skipif(sys.platform != "linux", reason="linux-only"),
-        ),
-        pytest.param(
-            r"C:\Program Files\Google\Chrome\Application\chrome.exe",
-            id="windows",
-            marks=pytest.mark.skipif(sys.platform != "win32", reason="windows-only"),
-        ),
-    ),
-)
-@mock.patch.object(browsers, "_launch")
-def test_launch(mock_launch: mock.MagicMock, chrome_path: str) -> None:
-    browsers.launch("chrome", url="https://github.com/roniemartinez/browsers")
-    mock_launch.assert_called_with("chrome", chrome_path, "https://github.com/roniemartinez/browsers", [])
-
-
-@mock.patch.object(browsers, "_launch")
-def test_launch_no_browser(mock_launch: mock.MagicMock) -> None:
-    browsers.launch("hello", url="https://github.com/roniemartinez/browsers")
-    mock_launch.assert_not_called()
